update ANU world rankings
use QS again, because that's where ANU's number one

Obviously anyone with their head screwed on (hi tenure committee!) knows that
this sort of "a-la-carte ranking selection" is truly bullshit (in the Graeberian
sense). But it's a facade we mustn't let slip.

Let this commit message stand as silent protest against this practice, even as I
make the change to my CV that perpetuates it.