add announcement blog post
This is a bit of an experiment to see who looks at my blog source (rather than
just the published version on the web). Oh the secrets you might find as you dig
through the source (and don't forget about the commit messages).

Anyway, this news will be public sooner or later, and I'm happy for you to share
this post far & wide as my attempt to communicate all the feelings I have about
my upcoming move to SoCy.

Happy trails.