update affiliation to ANU School of Cybernetics
Today's my first day in the new School of Cybernetics. So exciting! A big
thankyou to all the folks at the School of Computing who have supported me & my
team over the last decade.

If you want to hear a bit more about what this move means, I wrote a blog post
about it:

https://benswift.me/blog/2021/05/11/moving-to-the-school-of-cybernetics/

If this is your first time reading a git commit message, then welcome! Not sure
whether McLuhan would say that this is a hot or cold medium---and it's not
traditionally how people announce important job news---but given that my blog is
open source it felt apposite :)