update highlight js file
I've been doing more work in Elixir recently, but the hljs I use for this blog
wasn't "compiled" with the Elixir grammar.

To fix this, I generated a new custom build of hljs from
<https://highlightjs.org/download/>, then renamed the downloaded
`highlight.min.js` to `highlight.pack.js`, and copied it over the top of the one
currently in `assets/js/` in this repo.

seems to work :shrug: